rename progress_config.rs to progress.rs
