+++ src/plater/execution_mode/single_threaded_runner.rs
@@ -1,7 +1,7 @@
 use std::time::Duration;
 
 use crate::plater::placer::{Placer, N};
-use crate::plater::progress_config::{ProgressMessage, ProgressMessenger};
+use crate::plater::progress::{ProgressMessage, ProgressMessenger};
 use crate::plater::recommender::{Recommender, Suggestion};
 use crate::plater::request::{PlacingError, Request};
 use crate::plater::solution::{get_smallest_solution, Solution};
